v3.1.1
Version 3.1.1 - Many bugs fixed, improved alignment, option to load location from profile, option to crop image edges.

+++ hops/application_5_photometry.py
@@ -157,7 +157,7 @@ class PhotometryWindow(MainWindow):
             except:
                 self.targets_y_position.append(self.Label(text=0.0, instance=float))
 
-            self.targets_peak.append(self.Label(text=0.0, instance=float))
+            self.targets_peak.append(self.Label(text=0.0, instance=int))
             self.targets_max_hwhm.append(self.Label(text=0.0, instance=float))
             self.targets_total_flux.append(self.Label(text=0.0, instance=float))
             try:
